Remove SQLite: PostgreSQL-only database layer
- Remove all SQLite code paths from Database class — psycopg2 is now a
  required dependency, not optional
- Remove sqlite3 import, is_postgres flag, lastrowid property
- Use RETURNING id for INSERT operations that need the new row ID
- Fix verification_timestamp column: INTEGER → BIGINT for millisecond
  timestamps
- All tests now run against a local PostgreSQL container
  (podman run postgres:16-alpine on port 5433)
- Add conftest.py autouse fixture to set PULSE_AGENT_DATABASE_URL
- Update all test fixtures to TRUNCATE with RESTART IDENTITY for isolation
- Migrate eval outcome tests from sqlite3 to PostgreSQL

+++ sre_agent/db_schema.py
@@ -83,7 +83,7 @@ CREATE TABLE IF NOT EXISTS actions (
     resources TEXT,
     verification_status TEXT,
     verification_evidence TEXT,
-    verification_timestamp INTEGER
+    verification_timestamp BIGINT
 );
 """
